Extract SongQueue into a separate module
Moved the SongQueue and BgDownloadSongQueue classes from music_service.py to a new dedicated file, song_queue.py, for better modularity and code organization. Updated relevant imports in dependent modules to reflect this change.
